requirements.txt
discord==2.2.2
Pillow==9.4.0
PyAutoGUI==0.9.53
pynput==1.7.6

